Analyzing companies data in SQL
In this project I will analyze data for companies. A dataset containing information about each company on the INC 5000 list in 2019.
Fields include company name, industry, year of incorporation, website and location, as well as 2019 revenue, percentage growth, number of employees (annual growth) and number of years on the list.